+++ src/main/java/insane96mcp/iguanatweaksexpanded/data/generator/ISEGlobalLootModifierProvider.java
@@ -1,6 +1,5 @@
 package insane96mcp.iguanatweaksexpanded.data.generator;
 
-import insane96mcp.iguanatweaksexpanded.module.experience.enchanting.EnchantingFeature;
 import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
 import net.minecraft.data.PackOutput;
 import net.minecraftforge.common.data.GlobalLootModifierProvider;
@@ -12,7 +11,6 @@ public class ISEGlobalLootModifierProvider extends GlobalLootModifierProvider {
 
     @Override
     protected void start() {
-        EnchantingFeature.addGlobalLoot(this);
         NewEnchantmentsFeature.addGlobalLoot(this);
     }
 }