+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Upcoming
+### Changes
+* Removed cleansed and enchanted lapis
+  * You can now purify items with bottle o' enchanting, which is the same as the old infusing, with +1 to enchantability and over level enchantments
+* Durium, Solarium, Keego, Quaron, Soul Steel, Repair Kits and Forging are now disabled by default
+
 ## 1.24.5
 ### Additions
 * Tide's fishing rod has now the same enchantability as vanilla's